requirements.txt
numpy
opencv-python-headless
PyYAML
tqdm